+++ src/main/java/com/jeffpeng/jmod/primitives/JMODPluginInfo.java
@@ -1,7 +1,6 @@
 package com.jeffpeng.jmod.primitives;
 
 import java.util.List;
-import java.util.Map;
 
 public class JMODPluginInfo {
 	public String pluginid;
@@ -11,6 +10,8 @@ public class JMODPluginInfo {
 	public List<String> authors;
 	public String url;
 	public String credits;
+	public String archivebase;
 	public Object scriptingobjects;
+	public List<String> classtransformers;
 
 }