+++ src/main/java/com/jeffpeng/jmod/primitives/OwnedObject.java
@@ -1,8 +1,9 @@
 package com.jeffpeng.jmod.primitives;
 
+import java.util.Map;
+
 import org.apache.logging.log4j.Logger;
 
-import com.jeffpeng.jmod.Config;
 import com.jeffpeng.jmod.JMODRepresentation;
 import com.jeffpeng.jmod.Lib;
 import com.jeffpeng.jmod.interfaces.IEventObject;
@@ -11,7 +12,7 @@ import com.jeffpeng.jmod.interfaces.IOwned;
 
 public class OwnedObject implements IEventObject, IOwned {
 	protected JMODRepresentation owner;
-	protected Config config;
+	protected Map<String,Object> config;
 	protected Lib lib;
 	protected Logger log;