+++ src/main/java/com/jeffpeng/jmod/scripting/Game.java
@@ -1,7 +1,6 @@
 package com.jeffpeng.jmod.scripting;
 
 import com.jeffpeng.jmod.JMODRepresentation;
-import com.jeffpeng.jmod.interfaces.IEventObject;
 import com.jeffpeng.jmod.primitives.OwnedObject;
 
 public class Game extends OwnedObject {