+++ src/main/java/com/jeffpeng/jmod/scripting/ScriptObject.java
@@ -1,27 +1,42 @@
 package com.jeffpeng.jmod.scripting;
 
-import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
-
-import javax.script.Bindings;
-
+import java.util.HashMap;
 import net.minecraft.block.Block;
 import net.minecraft.item.ItemStack;
 
-import com.jeffpeng.jmod.JMOD;
 import com.jeffpeng.jmod.JMODLoader;
 import com.jeffpeng.jmod.Lib;
-import com.jeffpeng.jmod.actions.*;
-import com.jeffpeng.jmod.descriptors.*;
+import com.jeffpeng.jmod.actions.AddArmorMaterial;
+import com.jeffpeng.jmod.actions.AddBlock;
+import com.jeffpeng.jmod.actions.AddBlockDrop;
+import com.jeffpeng.jmod.actions.AddChestLoot;
+import com.jeffpeng.jmod.actions.AddCreativeTab;
+import com.jeffpeng.jmod.actions.AddFluid;
+import com.jeffpeng.jmod.actions.AddItem;
+import com.jeffpeng.jmod.actions.AddOreDictionaryEntry;
+import com.jeffpeng.jmod.actions.AddOreGeneration;
+import com.jeffpeng.jmod.actions.AddShapedRecipe;
+import com.jeffpeng.jmod.actions.AddShapelessRecipe;
+import com.jeffpeng.jmod.actions.AddSmeltingRecipe;
+import com.jeffpeng.jmod.actions.AddStartingPlatform;
+import com.jeffpeng.jmod.actions.AddToolMaterial;
+import com.jeffpeng.jmod.actions.RemoveChestLoot;
+import com.jeffpeng.jmod.actions.RemoveOreDictionaryEntry;
+import com.jeffpeng.jmod.actions.RemoveRecipe;
+import com.jeffpeng.jmod.actions.RemoveSmeltingRecipe;
+import com.jeffpeng.jmod.actions.SetBlockProperties;
+import com.jeffpeng.jmod.descriptors.ArmorDataDescriptor;
+import com.jeffpeng.jmod.descriptors.ColorDescriptor;
+import com.jeffpeng.jmod.descriptors.FoodDataDescriptor;
+import com.jeffpeng.jmod.descriptors.ItemStackSubstituteDescriptor;
+import com.jeffpeng.jmod.descriptors.ToolDataDescriptor;
+import com.jeffpeng.jmod.descriptors.TooltipDescriptor;
 import com.jeffpeng.jmod.primitives.OwnedObject;
-import com.jeffpeng.jmod.scripting.mods.*;
 import com.jeffpeng.jmod.util.JSFunctionWrapper;
 import com.jeffpeng.jmod.validator.Validator;
-
+@SuppressWarnings("unchecked")
 public class ScriptObject extends OwnedObject {
 	
 	private JScript jscriptInstance;
@@ -38,11 +53,6 @@ public class ScriptObject extends OwnedObject {
 	public World World = new World(owner);
 	public Players Players = new Players(owner);
 	public Game Game = new Game(owner);
-	
-	
-	public Chisel Chisel = new Chisel(owner);
-	public Applecore Applecore = new Applecore(owner);
-	public Sync Sync = new Sync(owner);
 		
 	public void loadjs(String script){
 		jscriptInstance.evalScript(script);
@@ -84,11 +94,11 @@ public class ScriptObject extends OwnedObject {
 	
 	
 	public void addMetalIngot(String name){
-		config.metalingots.add(name);
+		((ArrayList<String>)config.get("metalingots")).add(name);
 	}
 	
 	public void addMetalBlock(String name){
-		config.metalblocks.add(name);
+		((ArrayList<String>)config.get("metalblocks")).add(name);
 	}
 	
 	public  AddToolMaterial addToolMaterial(String name,int harvestLevel, int durability, float efficiency, float damage, int enchantability, String repairmaterial){
@@ -107,33 +117,21 @@ public class ScriptObject extends OwnedObject {
 		new RemoveSmeltingRecipe(owner,result);
 	}
 	
-	public void hideFromNEI(String target){
-		if(owner.testForMod("NotEnoughItems")) new HideNEIItem(owner,target);
-	}
+
 	
 	public  TooltipDescriptor addToolTip(String[] target, String[] lines){
 		TooltipDescriptor newTooltip = new TooltipDescriptor(target,lines);
-		config.tooltips.add(newTooltip);
+		((ArrayList<TooltipDescriptor>)config.get("tooltips")).add(newTooltip);
 		return newTooltip;
 	}
 	
 	public  ItemStackSubstituteDescriptor itemStackSubstitute(String source, String target){
 		ItemStackSubstituteDescriptor newISD = new ItemStackSubstituteDescriptor(source,target);
-		config.itemstacksubstitutes.add(newISD);
+		((ArrayList<ItemStackSubstituteDescriptor>)config.get("itemstacksubstitutes")).add(newISD);
 		return newISD;
 	}
 	
-	public  AlloyDescriptor addAlloy(String result, String one, String two, int amount){
-		AlloyDescriptor newAD = new AlloyDescriptor(result,one,two,amount);
-		config.alloymap.add(newAD);
-		return newAD;
-	}
-	
-	public  AlloyDescriptor addAlloy(String result, String one, int amount){
-		AlloyDescriptor newAD = new AlloyDescriptor(result,one,amount);
-		config.alloymap.add(newAD);
-		return newAD;
-	}
+
 	
 	public void addOreDict(String is, String entry){
 		new AddOreDictionaryEntry(owner,is,entry);
@@ -152,12 +150,12 @@ public class ScriptObject extends OwnedObject {
 	}
 	
 	public  void dependency(String modid, String name){
-		config.moddependencies.put(modid, name);
+		((HashMap<String,String>)config.get("moddependencies")).put(modid, name);
 	}
 	
 	public  ColorDescriptor defineColor(String color, int red, int green, int blue){
 		ColorDescriptor newColor = new ColorDescriptor(red,green,blue);
-		config.colors.put(color, newColor);
+		((HashMap<String,ColorDescriptor>)config.get("colors")).put(color, newColor);
 		return newColor;
 	}
 	
@@ -179,7 +177,7 @@ public class ScriptObject extends OwnedObject {
 	
 	public AddBlockDrop addBlockDrop(String block, String itemstack, Integer chance, boolean stopeventchain, boolean playeronly){
 		AddBlockDrop newBDD = new AddBlockDrop(owner,block, itemstack,chance, stopeventchain,playeronly);
-		config.blockDrops.add(newBDD);
+		((ArrayList<AddBlockDrop>) config.get("blockDrops")).add(newBDD);
 		return newBDD;
 	}